fixup: share styleJson via symlink
