+++ projects/maps-ng/src/components/si-map/shared/themes/map-style.ts
@@ -1,0 +1,1 @@
+../../../../../../element-ng/maplibre/map-style.ts
